+++ static/js/apps/event/main.ts
@@ -49,7 +49,9 @@ function renderPage(): void {
   // Event place
   const placeDcid = document.getElementById("place").dataset.dcid;
   const placeName = document.getElementById("place").dataset.name || placeDcid;
-  const placeTypes = [document.getElementById("place").dataset.types] || [];
+  const placeTypes =
+    JSON.parse(document.getElementById("place").dataset.types) || [];
+  console.log(placeTypes);
   const place = { dcid: placeDcid, name: placeName, types: placeTypes };
   const parentPlaces = JSON.parse(
     document.getElementById("place").dataset.parents || "[]"
