+++ static/js/place/chart_embed.tsx
@@ -19,7 +19,10 @@ import React from "react";
 import { Button, Modal, ModalBody, ModalFooter, ModalHeader } from "reactstrap";
 
 import { wrap } from "../chart/base";
-import { CLASS_DC_CHART_EXIST } from "../constants/css_constants";
+import {
+  CLASS_DC_CHART_EXIST,
+  CLASS_DC_CHART_HOLDER,
+} from "../constants/css_constants";
 import { intl } from "../i18n/i18n";
 import { randDomId, saveToFile, urlToDomain } from "../shared/util";
 
@@ -192,8 +195,6 @@ class ChartEmbed extends React.Component<unknown, ChartEmbedStateType> {
         sourcesHeight +
         SOURCES_MARGIN
     );
-    svg.attr("class", CLASS_DC_CHART_EXIST);
-
     const s = new XMLSerializer();
     const svgXml = s.serializeToString(svg.node());
     container.innerHTML = "";
@@ -301,6 +302,7 @@ class ChartEmbed extends React.Component<unknown, ChartEmbedStateType> {
         "data:image/svg+xml," + encodeURIComponent(chartDownloadXml);
       imageElement.src = chartBase64;
       this.svgContainerElement.current.append(imageElement);
+      imageElement.className = CLASS_DC_CHART_EXIST;
       this.setState({ chartDownloadXml });
     }
 
@@ -310,6 +312,7 @@ class ChartEmbed extends React.Component<unknown, ChartEmbedStateType> {
       const chartBase64 =
         "data:image/svg+xml," + encodeURIComponent(chartDownloadXml);
       imageElement.src = chartBase64;
+      imageElement.className = CLASS_DC_CHART_EXIST;
       this.svgContainerElement.current.append(imageElement);
       this.setState({ chartDownloadXml });
     }
@@ -360,7 +363,7 @@ class ChartEmbed extends React.Component<unknown, ChartEmbedStateType> {
         <ModalBody>
           <div
             ref={this.svgContainerElement}
-            className="modal-chart-container"
+            className={`modal-chart-container ${CLASS_DC_CHART_HOLDER}`}
           ></div>
           <textarea
             className="copy-svg mt-3"
